ponderada_6_7/src/consumer_service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
confluent-kafka==2.3.0
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
idna==3.6
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
